Refactor CSS styles and update scripts in package.json
This commit removes individual CSS files and combines them into a single style.css file to improve code organization. The scripts in package.json are also updated to adopt npm-run-all for parallel execution. Additionally, prettier is added to backend package.json for linting, "backend/.gitignore" is added to exclude node_modules, and an unused SVG file is deleted. A proper table display for posts is implemented in App.tsx. Lastly, minor adjustments are made to the mock database 'db.json.'